refactor : split service into readOnly and writeonly in member
